fix -R, --revert-settings to only remove the preferences file
MuseScore*.ini, and not also shortcuts, custom workspaces and palettes,
translations, etc., this is left for -F --factory-settings,
see https://musescore.org/en/node/276560

fix deletePreferences to only delete dataPath on useFactorySettings

+++ mscore/musescore.cpp
@@ -6697,7 +6697,8 @@ int main(int argc, char* av[])
             dataPath = QStandardPaths::writableLocation(QStandardPaths::DataLocation);
 
       if (deletePreferences) {
-            QDir(dataPath).removeRecursively();
+            if (useFactorySettings)
+                  QDir(dataPath).removeRecursively();
             QSettings settings;
             QFile::remove(settings.fileName() + ".lock"); //forcibly remove lock
             QFile::remove(settings.fileName());
